requirements.txt
fastapi
uvicorn
azure-ai-projects
azure-identity
streamlit
python-dotenv